Add docs for the block system
