+++ src/client/world/chunk_mesh_generation.h
@@ -5,7 +5,7 @@
 class Chunk;
 
 #include "client_voxel.h"
-#include <common/world/voxel_types.h>
+#include <common/world/voxel_registry.h>
 
 ChunkMesh makeChunkMesh(const Chunk &chunk,
                         const VoxelRegistry<ClientVoxel> &voxelData);
