+++ src/server/network/server.h
@@ -5,6 +5,7 @@
 #include <array>
 #include <common/network/net_host.h>
 #include <common/world/chunk_manager.h>
+#include <common/world/voxel_registry.h>
 #include <glm/gtc/matrix_transform.hpp>
 #include <queue>
 #include <unordered_map>
